requirements.txt
setuptools<70  # pkg_resources for nlp_dedup (Python 3.14+)
nlp_dedup>=0.1.0
ollama>=0.3.0
mlx>=0.22.0
mlx-lm[train]>=0.22.0
datasets>=2.18.0
huggingface_hub>=0.23.0
transformers>=4.36.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
rich>=13.0.0
anthropic>=0.39.0
python-dotenv>=1.0.0
modal>=0.64.0
